+ additional comments about refactoring removal targets + moved policy_data.py
